worker/pyproject.toml
[project]
name = "douyin-worker"
version = "0.1.0"
description = "Celery worker for processing Douyin videos"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "celery[redis]>=5.3.6",
    "redis>=5.0.1",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.0",
]

[project.scripts]
douyin-cli = "cli:main"

[tool.uv]
dev-dependencies = [
    "flower>=2.0.0",
]
